[uontt] get details working properly; tidy up
getting details working correctly was subtle (at least to me!).
`casperjs` creates a stack of operations, starting with `.start()`,
continuing with `.then()` invocations, and begins execution with `.run()`
finally. commonly in a `.then()` one executes `.evaluate()`, which
executes some code within the context of the remote DOM. values are
returned from the `@evaluate()` without any problem __BUT__ values cannot
be passed between `@then()` steps directly. instead one must assign
values to globals so that they are available in subsequent steps.
this is particularly true if, as here, subsequent steps are dynamically
computed, and added to the stack at a later point.

in particular, `if do_details` introduces the inclusion of module details
as a `@then()` step. the first thing the step must do is capture the existing
`tts` as `_tts` to iterate over, before zeroing `tts` to capture the result of the
iteration. it then iterates over `_tts`:
+ construct the URL;
+ stack a `@then()` step to open and fetch the contents of the URL;
+ stack a `@then()` step to evaluate the results once the fetch is complete.

the result of the last step is to return an entry which is pushed into `tts`.
the end result is data captured in `tts`.